Pos compare
Bench 13253579

+++ 4k.c
@@ -274,9 +274,9 @@ typedef struct [[nodiscard]] {
 
 [[nodiscard]] static bool position_equal(const Position *const restrict lhs,
                                          const Position *const restrict rhs) {
-  static_assert(sizeof(Position) % sizeof(u32) == 0);
-  for (u32 i = 0; i < sizeof(Position) / sizeof(u32); i++) {
-    if (((const u32 *)lhs)[i] != ((const u32 *)rhs)[i]) {
+  static_assert(sizeof(Position) % sizeof(u64) == 0);
+  for (u32 i = 0; i < sizeof(Position) / sizeof(u64); i++) {
+    if (((const u64 *)lhs)[i] != ((const u64 *)rhs)[i]) {
       return false;
     }
   }
